tar: move tar under root level namespace
